Fix JC-556: 1. Move 'acceptType' and 'minSize' mat-errors to item-manager component 2. Update 'wrongImage' mat-error in anime-form component 3. Rename 'dates-validator' util to 'anime-managers-form-validate' 4. Add 'maxLengthValidator' to 'anime-managers-form-validate.utils'
