--- apps/angular/src/core/utils/dates-validate.utils.ts
@@ -1,11 +1,0 @@
-import { ValidatorFn, AbstractControl, ValidationErrors } from '@angular/forms';
-
-/** Checks if end date is after start date. */
-export function datesValidator(): ValidatorFn {
-	return (control: AbstractControl): ValidationErrors | null => {
-		if (control.value >= control.parent?.get('airedStart')?.value) {
-			return null;
-		}
-		return { matching: true };
-	};
-}
